spark/requirements.txt
path
configparser
pyspark
numpy
mysql-connector-python
pandas
loguru
requests